Restore `sortBy` option

+++ src/lib/database.ts
@@ -1,17 +1,19 @@
-import { openDB } from 'idb/with-async-ittr';
+import { openDB } from 'idb';
+
+import { updateGitHubStars } from './github';
 
 import type { GitHubRepo } from './github';
-import type { DBSchema, IDBPDatabase } from 'idb/with-async-ittr';
+import type { DBSchema, IDBPDatabase } from 'idb';
 
-const DB_VERSION = 10;
+const DB_VERSION = 17;
 
 export interface TargetsDatabase extends DBSchema {
 	starredGitHubRepos: {
-		key: GitHubRepo['name'];
+		key: [GitHubRepo['owner'], GitHubRepo['name']];
 		// eslint-disable-next-line id-denylist -- Name imposed by the base interface
 		value: GitHubRepo;
 		indexes: {
-			name: string;
+			starredAt: string;
 		};
 	};
 }
@@ -19,9 +21,14 @@ export interface TargetsDatabase extends DBSchema {
 export function getDatabase(): Promise<IDBPDatabase<TargetsDatabase>> {
 	return openDB<TargetsDatabase>('targets', DB_VERSION, {
 		upgrade(database) {
-			database.deleteObjectStore('starredGitHubRepos');
-			const store = database.createObjectStore('starredGitHubRepos', { keyPath: ['owner', 'name'] });
-			store.createIndex('name', 'name');
+			if (database.objectStoreNames.contains('starredGitHubRepos')) {
+				database.deleteObjectStore('starredGitHubRepos');
+			}
+
+			const store = database.createObjectStore('starredGitHubRepos', { keyPath: ['name', 'owner'] });
+			store.createIndex('starredAt', 'starredAt');
+
+			void updateGitHubStars();
 		},
 	});
 }
